requirements.txt
streamlit==1.28.1
openai==1.3.7
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
